requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
pymupdf
gunicorn
protobuf
langchain
groq
python-dotenv
pymupdf
